Added parent-child corrections option
