+++ packages/go-enrich/src/go_enrich/utils.py
@@ -1,5 +1,5 @@
-from typing import Iterator, Tuple
 from pathlib import Path
+from typing import Dict, Iterator, List, Tuple
 
 
 def target_genes_file(file_path: Path) -> Iterator[str]:
@@ -49,3 +49,33 @@ def gene_to_go_term_mapping_file(
                 )
 
             yield (node_pair[0], node_pair[1])
+
+
+def benjamini_hochberg_fdr(
+    node_p_values: Dict[str, float], fdr: float = 0.01
+) -> List[str]:
+    ranks: Dict[str, float] = {
+        key: rank
+        for (rank, key) in enumerate(
+            sorted(node_p_values, key=lambda x: node_p_values[x]), start=1
+        )
+    }
+
+    imq = {
+        key: (ranks[key] / len(node_p_values)) * fdr
+        for key in node_p_values
+    }
+
+    smaller_pvalues = dict(
+        filter(lambda item: item[1] < imq[item[0]], node_p_values.items())
+    )
+
+    # max_key, max_value = max(smaller_pvalues.items(), key=lambda x: x[1])
+    max_key = max(smaller_pvalues.keys(), key=lambda x: node_p_values[x])
+
+    max_rank = ranks[max_key]
+
+    return sorted(
+        (x for x in node_p_values if ranks[x] <= max_rank),
+        key=lambda x: node_p_values[x],
+    )
